Create RNG deathrattle class
